+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+### Added
+* [Demo] Fix for Mobile SDK Gamma Canary failures in saucelabs. Added support for overridden meeting endpoint url in demo app.
+
 ## [0.19.0] - 2022-02-24
 
 ### Added